requirements.txt
clearbit==0.1.7
django==2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
requests==2.21.0
